Styles and CatergoryView restructured
