+++ src/components/CategoryView/CategoryView.styles.ts
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+export const DeleteButton = styled.button`
+    font-size: 1em;
+    margin: 1em;
+    margin-left:0em;
+    padding: 0.25em 1em;
+    border-radius: 3px;
+    background-color: white;
+    color: tomato; 
+    border: 2px solid tomato;
+`;
+
+export const AddButton = styled.button`
+    font-size: 1em;
+    margin: 1em;
+    padding: 0.25em 1em;
+    border-radius: 3px;
+    background-color: white;
+    color: mediumseagreen;
+    border: 2px solid mediumseagreen;
+`;
